requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
dj-database-url==2.2.0
gunicorn==22.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
whitenoise==6.7.0
